Code reorganization/cleanup: extract dialogs and associated logic from the export/publish provider
