+++ PhotoDeckPublishServiceProvider.lua
@@ -1,21 +1,14 @@
 local LrApplication = import 'LrApplication'
-local LrBinding = import 'LrBinding'
-local LrDialogs = import 'LrDialogs'
 local LrFileUtils = import 'LrFileUtils'
 local LrTasks = import 'LrTasks'
-local LrView = import 'LrView'
 local LrErrors = import 'LrErrors'
 local LrHttp = import 'LrHttp'
 
 local logger = import 'LrLogger'( 'PhotoDeckPublishLightroomPlugin' )
-
 logger:enable('logfile')
 
 local PhotoDeckAPI = require 'PhotoDeckAPI'
-local PhotoDeckUtils = require 'PhotoDeckUtils'
-local printTable = PhotoDeckUtils.printTable
-local filter = PhotoDeckUtils.filter
-local map = PhotoDeckUtils.map
+local PhotoDeckDialogs = require 'PhotoDeckDialogs'
 
 local publishServiceProvider = {}
 
@@ -23,6 +16,7 @@ local publishServiceProvider = {}
 publishServiceProvider.hideSections = { 'exportLocation' }
 publishServiceProvider.allowFileFormats = { 'JPEG', 'TIFF', 'Original' }
 publishServiceProvider.hidePrintResolution = true
+
 
 -- General plugin configuration for publish operations
 publishServiceProvider.supportsIncrementalPublish = true
@@ -32,7 +26,31 @@ publishServiceProvider.titleForPublishedCollectionSet = LOC("$$$/PhotoDeck/Publi
 publishServiceProvider.titleForGoToPublishedCollection = LOC("$$$/PhotoDeck/Publish/GoToCollection=Go to PhotoDeck Gallery")
 publishServiceProvider.titleForGoToPublishedPhoto =  LOC("$$$/PhotoDeck/Publish/GoToPhoto=Go to Photo in PhotoDeck Gallery")
 
--- these fields get stored between uses
+
+-- Photo metadata changes that should retrigger a republish
+publishServiceProvider.metadataThatTriggersRepublish = function(publishSettings)
+  return {
+    default = false,
+    rating = true,
+    title = true,
+    caption = true,
+    creator = true,
+    keywords = true,
+    dateCreated = true,
+    headline = true,
+    iptcSubjectCode = true,
+    dateCreated = true,
+    location = true,
+    city = true,
+    stateProvince = true,
+    country = true,
+    copyright = true,
+    customMetadata = false
+  }
+end
+
+
+-- These fields get stored between uses
 publishServiceProvider.exportPresetFields = {
   { key = 'username', default = "" },
   { key = 'password', default = "" },
@@ -42,396 +60,26 @@ publishServiceProvider.exportPresetFields = {
   { key = 'uploadOnRepublish', default = false }
 }
 
-local function  updateApiKeyAndSecret(propertyTable)
-  local f = LrView.osFactory()
-  local c = f:column {
-    bind_to_object = propertyTable,
-    spacing = f:dialog_spacing(),
-    f:row {
-      f:static_text {
-        title = LOC "$$$/PhotoDeck/ApiKeyDialog/ApiKey=API Key:",
-        width = LrView.share "label_width",
-        alignment = "right",
-      },
-      f:edit_field {
-        value = LrView.bind 'apiKey',
-        immediate = false,
-        width_in_chars = 40,
-      }
-    },
-    f:row {
-      f:static_text {
-        title = LOC "$$$/PhotoDeck/ApiKeyDialog/ApiSecret=API Secret:",
-        width = LrView.share "label_width",
-        alignment = "right",
-      },
-      f:edit_field {
-        value = LrView.bind 'apiSecret',
-        immediate = false,
-        width_in_chars = 40,
-      },
-    },
-  }
-  local result = LrDialogs.presentModalDialog({
-    title = LOC "$$$/PhotoDeck/ApiKeyDialog/Title=PhotoDeck API Keys",
-    contents = c,
-  })
-  return propertyTable
+
+-- Plugin settings dialog
+publishServiceProvider.startDialog = PhotoDeckDialogs.startDialog
+publishServiceProvider.sectionsForTopOfDialog = PhotoDeckDialogs.sectionsForTopOfDialog
+
+
+-- Published collection / collection set settings dialog
+publishServiceProvider.viewForCollectionSettings = PhotoDeckDialogs.viewForCollectionSettings
+publishServiceProvider.viewForCollectionSetSettings = PhotoDeckDialogs.viewForCollectionSettings
+
+
+-- Published collection name validation
+publishServiceProvider.validatePublishedCollectionName = function(proposedName)
+  -- string needs to be valid UTF-8 (3 multibyte chars max) and less than 200 bytes
+  local length = string.len(proposedName)
+  return length > 0 and length < 200
 end
 
-local function chooseWebsite(propertyTable)
-  local f = LrView.osFactory()
-  local c = f:row {
-    spacing = f:dialog_spacing(),
-    bind_to_object = propertyTable,
-    f:popup_menu {
-      items = LrView.bind 'websiteChoices',
-      value = LrView.bind 'websiteChosen',
-    },
-  }
-  local result = LrDialogs.presentModalDialog({
-    title = LOC "$$$/PhotoDeck/WebsitesDialog/Title=PhotoDeck Websites",
-    contents = c,
-  })
-  return propertyTable
-end
 
-local function updateCantExportBecause(propertyTable)
-  if not propertyTable.loggedin then
-    propertyTable.LR_cantExportBecause = LOC "$$$/PhotoDeck/AccountDialog/NoLogin=You haven't logged in to PhotoDeck yet."
-    return
-  end
-  propertyTable.LR_cantExportBecause = nil
-end
-
-local function onWebsiteSelect(propertyTable, key, value)
-  propertyTable.websiteName = propertyTable.websites[value].title
-end
-
-local function chooseGalleryDisplayStyle(propertyTable, collectionInfo)
-  local f = LrView.osFactory()
-  local c = f:row {
-    spacing = f:dialog_spacing(),
-    bind_to_object = collectionInfo,
-    f:popup_menu {
-      items = LrView.bind 'galleryDisplayStyles',
-      value = LrView.bind { bind_to_object = collectionInfo, key = 'display_style' }
-    },
-  }
-  local result = LrDialogs.presentModalDialog({
-    title = LOC "$$$/PhotoDeck/GalleryDisplayStylesDialog/Title=Gallery Display Styles",
-    contents = c,
-  })
-  return propertyTable
-end
-
-local function ping(propertyTable)
-  propertyTable.connectionStatus = LOC "$$$/PhotoDeck/ConnectionStatus/Connecting=Connecting to PhotoDeck^."
-  LrTasks.startAsyncTask(function()
-    PhotoDeckAPI.connect(propertyTable.apiKey, propertyTable.apiSecret, propertyTable.username, propertyTable.password)
-    local ping, error_msg = PhotoDeckAPI.ping()
-    if error_msg then
-      propertyTable.connectionStatus = LOC("$$$/PhotoDeck/ConnectionStatus/Failed=Connection failed: ^1", error_msg)
-    elseif propertyTable.loggedin then
-      propertyTable.connectionStatus = LOC "$$$/PhotoDeck/ConnectionStatus/Connected=Connected"
-    else
-      propertyTable.connectionStatus = LOC "$$$/PhotoDeck/ConnectionStatus/ConnectedPleaseLogin=Connected, please log in"
-    end
-  end, 'PhotoDeckAPI Ping')
-end
-
-local function login(propertyTable)
-  propertyTable.connectionStatus = LOC "$$$/PhotoDeck/ConnectionStatus/LoggingIn=Logging in^."
-  LrTasks.startAsyncTask(function()
-    PhotoDeckAPI.connect(propertyTable.apiKey, propertyTable.apiSecret, propertyTable.username, propertyTable.password)
-    local result, error_msg = PhotoDeckAPI.whoami()
-    if error_msg then
-      propertyTable.connectionStatus = error_msg
-    elseif not PhotoDeckAPI.loggedin then
-      propertyTable.connectionStatus = LOC "$$$/PhotoDeck/ConnectionStatus/CredentialsError=Couldn't log in using those credentials"
-    else
-      propertyTable.connectionStatus = LOC("$$$/PhotoDeck/ConnectionStatus/LoggedInAs=Logged in as ^1 ^2", result.firstname, result.lastname)
-    end
-    propertyTable.loggedin = PhotoDeckAPI.loggedin
-    
-    if PhotoDeckAPI.loggedin then
-      -- get available websites
-      websites, error_msg = PhotoDeckAPI.websites()
-      if error_msg then
-        propertyTable.connectionStatus = LOC("$$$/PhotoDeck/ConnectionStatus/FailedLoadingWebsite=Couldn't get your website: ^1", error_msg)
-        propertyTable.loggedin = PhotoDeckAPI.loggedin
-      else
-        propertyTable.websites = websites
-	propertyTable.websiteChoices = {}
-	local websitesCount = 0
-	local firstWebsite = nil
-	local foundCurrent = false
-        for k, v in pairs(propertyTable.websites) do
-	  websitesCount = websitesCount + 1
-          if not firstWebsite then
-	    firstWebsite = k
-          end
-	  if k == propertyTable.websiteChosen then
-	    foundCurrent = true
-	  end
-          table.insert(propertyTable.websiteChoices, { title = v.title .. " (" .. v.hostname .. ")", value = k })
-        end
-	if not foundCurrent then
-          -- automatically select first website
-	  propertyTable.websiteChosen = firstWebsite
-	end
-        propertyTable.multipleWebsites = websitesCount > 1
-        onWebsiteSelect(propertyTable, nil, propertyTable.websiteChosen)
-      end
-
-      -- show synchronization message if in progress
-      if not propertyTable.canSynchronize then
-        propertyTable.synchronizeGalleriesResult = LOC("$$$/PhotoDeck/SynchronizeStatus/InProgress=In progress^.")
-      end
-    end
-  end, 'PhotoDeckAPI Login')
-end
-
-local function synchronizeGalleries(propertyTable)
-  propertyTable.synchronizeGalleriesResult = LOC("$$$/PhotoDeck/SynchronizeStatus/Starting=Starting^.")
-  propertyTable.canSynchronize = false
-  LrTasks.startAsyncTask(function()
-    PhotoDeckAPI.connect(propertyTable.apiKey, propertyTable.apiSecret, propertyTable.username, propertyTable.password)
-    local result, error_msg = PhotoDeckAPI.synchronizeGalleries(propertyTable.websiteChosen, propertyTable)
-    if error_msg then
-      propertyTable.synchronizeGalleriesResult = error_msg
-    elseif result then
-      if result.created == 0 and result.deleted == 0 and result.updated == 0 and result.errors == 0 then
-        propertyTable.synchronizeGalleriesResult = LOC("$$$/PhotoDeck/SynchronizeStatus/FinishedWithNoChanges=Finished, no changes")
-      else
-        propertyTable.synchronizeGalleriesResult = LOC("$$$/PhotoDeck/SynchronizeStatus/FinishedWithChanges=Finished: ^1 created, ^2 deleted, ^3 updated, ^4 errors", result.created, result.deleted, result.updated, result.errors)
-      end
-    else
-      propertyTable.synchronizeGalleriesResult = "?"
-    end
-    propertyTable.canSynchronize = true
-  end, 'PhotoDeckAPI galleries synchronization')
-end
-
-local function onGalleryDisplayStyleSelect(propertyTable, key, value)
-  local chosenStyle = filter(propertyTable.galleryDisplayStyles, function(v) return v.value == value end)
-  if #chosenStyle > 0 then
-    propertyTable.galleryDisplayStyleName = chosenStyle[1].title
-  end
-end
-
-function publishServiceProvider.startDialog(propertyTable)
-  propertyTable.loggedin = false
-  propertyTable.websiteChoices = {}
-  propertyTable.galleryDisplayStyles = {}
-  propertyTable.multipleWebsites = false
-  propertyTable.websiteName = ''
-  propertyTable.canSynchronize = PhotoDeckAPI.canSynchronize
-  propertyTable.synchronizeGalleriesResult = ''
-
-  propertyTable:addObserver('loggedin', function() updateCantExportBecause(propertyTable) end)
-  updateCantExportBecause(propertyTable)
-
-  propertyTable:addObserver('websiteChosen', onWebsiteSelect)
-
-  local keysAreValid = PhotoDeckAPI.hasDistributionKeys or (
-    propertyTable.apiKey and propertyTable.apiKey ~= '' and
-    propertyTable.apiSecret and propertyTable.apiSecret ~= '')
-
-  if not keysAreValid then
-    propertyTable = updateApiKeyAndSecret(propertyTable)
-  end
-  if propertyTable.username and propertyTable.username ~= '' and
-     propertyTable.password and propertyTable.password ~= '' and
-     keysAreValid then
-    login(propertyTable)
-  else
-    ping(propertyTable)
-  end
-end
-
-function publishServiceProvider.sectionsForTopOfDialog( f, propertyTable )
-  -- LrMobdebug.on()
-  local isPublish = not not propertyTable.LR_publishService
-
-  local apiCredentials =  {
-    title = LOC "$$$/PhotoDeck/ApiKeyDialog/Title=PhotoDeck API Keys",
-    synopsis = LrView.bind 'connectionStatus',
-
-    f:row {
-      bind_to_object = propertyTable,
-      f:column {
-        f:row {
-          f:static_text {
-            title = LOC "$$$/PhotoDeck/ApiKeyDialog/ApiKey=API Key:",
-            width = LrView.share "label_width",
-            alignment = 'right'
-          },
-          f:static_text {
-            title = LrView.bind 'apiKey',
-            width_in_chars = 40,
-          }
-        },
-        f:row {
-          f:static_text {
-            title = LOC "$$$/PhotoDeck/ApiKeyDialog/ApiSecret=API Secret:",
-            width = LrView.share "label_width",
-            alignment = 'right'
-          },
-          f:static_text {
-            title = LrView.bind 'apiSecret',
-            width_in_chars = 40,
-          }
-        },
-      },
-      f:column {
-        f:push_button {
-          title = LOC "$$$/PhotoDeck/ApiKeyDialog/ChangeAction=Change",
-          action = function() propertyTable = updateApiKeyAndSecret(propertyTable) end,
-        },
-      },
-    },
-  }
-  local userAccount = {
-    title = LOC "$$$/PhotoDeck/AccountDialog/Title=PhotoDeck Account",
-    synopsis = LrView.bind 'connectionStatus',
-
-    f:row {
-      bind_to_object = propertyTable,
-      f:column {
-        f:row {
-          f:static_text {
-            title = LOC "$$$/PhotoDeck/AccountDialog/Email=Email:",
-            width = LrView.share "user_label_width",
-            alignment = 'right'
-          },
-          f:edit_field {
-            value = LrView.bind 'username',
-            width = 300,
-          }
-        },
-        f:row {
-          f:static_text {
-            title = LOC "$$$/PhotoDeck/AccountDialog/Password=Password:",
-            width = LrView.share "user_label_width",
-            alignment = 'right'
-          },
-          f:password_field {
-            value = LrView.bind 'password',
-            width = 300,
-          }
-        },
-      },
-
-      f:column {
-        f:push_button {
-          title = LOC "$$$/PhotoDeck/AccountDialog/LoginAction=Login",
-          enabled = LrBinding.negativeOfKey('loggedin'),
-          action = function () login(propertyTable) end
-        },
-      },
-    },
-
-    f:row {
-      bind_to_object = propertyTable,
-      f:column {
-        f:static_text {
-          title = "",
-          width = LrView.share "user_label_width",
-          alignment = 'right'
-        },
-      },
-      f:column {
-        f:static_text {
-          title = LrView.bind 'connectionStatus',
-	  font = '<system/small/bold>',
-	  width = 300,
-	  height_in_lines = 2
-        },
-      },
-    },
-
-    f:row {
-      bind_to_object = propertyTable,
-      f:column {
-	f:row {
-          f:static_text {
-	    visible = LrBinding.andAllKeys('LR_publishService', 'loggedin'),
-            title = LOC "$$$/PhotoDeck/AccountDialog/Website=Website:",
-            width = LrView.share "user_label_width",
-            alignment = 'right'
-          },
-          f:static_text {
-	    visible = LrBinding.andAllKeys('LR_publishService', 'loggedin'),
-            title = LrView.bind 'websiteName',
-            width = 300,
-          }
-        },
-      },
-
-      f:column {
-        f:push_button {
-          title = LOC "$$$/PhotoDeck/AccountDialog/WebsiteChangeAction=Change",
-	  visible = LrBinding.andAllKeys('LR_publishService', 'loggedin'),
-	  enabled = LrBinding.andAllKeys('loggedin', 'multipleWebsites'),
-          action = function() propertyTable = chooseWebsite(propertyTable) end,
-        },
-      },
-    }
-  }
-  local publishSettings = {
-    title = LOC "$$$/PhotoDeck/PublishOptionsDialog/Title=PhotoDeck Publish options",
-
-    f:row {
-      bind_to_object = propertyTable,
-
-      f:checkbox {
-        title = LOC "$$$/PhotoDeck/PublishOptionsDialog/UploadOnRepublish=Re-upload photo when re-publishing",
-        value = LrView.bind 'uploadOnRepublish'
-      }
-    },
-
-    f:row {
-      f:push_button {
-        title = LOC "$$$/PhotoDeck/PublishOptionsDialog/SynchronizeGalleriesAction=Import PhotoDeck galleries",
-        action = function()
-                   if not propertyTable.LR_publishService then
-		     -- publish service is not created yet (this is a new unsaved plugin instance)
-		     LrDialogs.message(LOC "$$$/PhotoDeck/PublishOptionsDialog/SaveFirst=Please save the settings first!")
-	           else
-		     local result = LrDialogs.confirm(
-		       LOC "$$$/PhotoDeck/PublishOptionsDialog/ConfirmTitle=This will mirror your existing PhotoDeck galleries in Lightroom.",
-		       LOC "$$$/PhotoDeck/PublishOptionsDialog/ConfirmSubtitle=Gallery content is currently not imported.",
-		       LOC "$$$/PhotoDeck/PublishOptionsDialog/ProceedAction=Proceed",
-		       LOC "$$$/PhotoDeck/PublishOptionsDialog/CancelAction=Cancel")
-		     if result == "ok" then
-                       synchronizeGalleries(propertyTable)
-		     end
-	           end
-	end,
-	enabled = LrBinding.andAllKeys('loggedin', 'canSynchronize'),
-      },
-
-      f:static_text {
-        title = LrView.bind 'synchronizeGalleriesResult',
-        alignment = 'right',
-        fill_horizontal = 1,
-        height_in_lines = 1,
-      },
-    }
-  }
-
-  local dialogs = {}
-  if not PhotoDeckAPI.hasDistributionKeys then
-    table.insert(dialogs, apiCredentials)
-  end
-  table.insert(dialogs, userAccount);
-  if isPublish then
-    table.insert(dialogs, publishSettings);
-  end
-  return dialogs
-end
-
+-- Process rendered photos (export or publish)
 local function getPhotoDeckPhotoIdsStoredInCatalog(photo)
   local str = photo:getPropertyForPlugin(_PLUGIN, "photoId")
   if not str then
@@ -638,10 +286,12 @@ function publishServiceProvider.processRenderedPhotos( functionContext, exportCo
   end
 
   progressScope:done()
-
 end
 
-publishServiceProvider.deletePhotosFromPublishedCollection = function( publishSettings, arrayOfPhotoIds, deletedCallback, localCollectionId )
+
+
+-- Delete photo from published collection
+publishServiceProvider.deletePhotosFromPublishedCollection = function(publishSettings, arrayOfPhotoIds, deletedCallback, localCollectionId)
   logger:trace('publishServiceProvider.deletePhotosFromPublishedCollection')
   PhotoDeckAPI.connect(publishSettings.apiKey, publishSettings.apiSecret, publishSettings.username, publishSettings.password)
   local catalog = LrApplication.activeCatalog()
@@ -688,130 +338,8 @@ publishServiceProvider.deletePhotosFromPublishedCollection = function( publishSe
   end
 end
 
-publishServiceProvider.validatePublishedCollectionName = function( proposedName )
-  -- string needs to be valid UTF-8 (3 multibyte chars max) and less than 200 bytes
-  local length = string.len(proposedName)
-  return length > 0 and length < 200
-end
 
-publishServiceProvider.metadataThatTriggersRepublish = function( publishSettings )
-  return {
-    default = false,
-    rating = true,
-    title = true,
-    caption = true,
-    creator = true,
-    keywords = true,
-    dateCreated = true,
-    headline = true,
-    iptcSubjectCode = true,
-    dateCreated = true,
-    location = true,
-    city = true,
-    stateProvince = true,
-    country = true,
-    copyright = true,
-    customMetadata = false
-  }
-end
-
-publishServiceProvider.viewForCollectionSettings = function( f, publishSettings, info )
-  info.collectionSettings:addObserver('display_style', onGalleryDisplayStyleSelect)
-
-  info.collectionSettings.galleryDisplayStyles = {}
-
-  publishSettings.connectionStatus = LOC "$$$/PhotoDeck/CollectionSettingsDialog/ConnectionStatus/Connecting=Connecting to PhotoDeck^."
-  LrTasks.startAsyncTask(function()
-    PhotoDeckAPI.connect(publishSettings.apiKey, publishSettings.apiSecret, publishSettings.username, publishSettings.password)
-    local error_msg = nil
-
-    local publishedCollection = info.publishedCollection
-    if publishedCollection then
-      local galleryId = publishedCollection:getRemoteId()
-      if galleryId and galleryID ~= '' then
-        -- read current live settings
-        local gallery
-        gallery, error_msg = PhotoDeckAPI.gallery(publishSettings.websiteChosen, galleryId)
-        if error_msg then
-          publishSettings.connectionStatus = LOC("$$$/PhotoDeck/CollectionSettingsDialog/ConnectionStatus/GalleryFailed=Error reading gallery from PhotoDeck: ^1", error_msg)
-        else
-	  info.collectionSettings.LR_liveName = gallery.name
-  	  info.collectionSettings.description = gallery.description
-	  info.collectionSettings.display_style = gallery.displaystyle
-        end
-      end
-    end
-
-    if not error_msg then
-      local galleryDisplayStyles
-      galleryDisplayStyles, error_msg = PhotoDeckAPI.galleryDisplayStyles(publishSettings.websiteChosen)
-      if error_msg then
-        publishSettings.connectionStatus = LOC("$$$/PhotoDeck/CollectionSettingsDialog/ConnectionStatus/GalleryDisplayStylesFailed=Error reading gallery display styles from PhotoDeck: ^1", error_msg)
-      elseif galleryDisplayStyles then
-        -- populate gallery display style choices
-        for k, v in pairs(galleryDisplayStyles) do
-          table.insert(info.collectionSettings.galleryDisplayStyles, { title = v.name, value = v.uuid })
-        end
-        if info.collectionSettings.display_style and info.collectionSettings.display_style ~= '' then
-          onGalleryDisplayStyleSelect(info.collectionSettings, nil, info.collectionSettings.display_style)
-        end
-      end
-    end
-
-    if not error_msg then
-      publishSettings.connectionStatus = LOC("$$$/PhotoDeck/CollectionSettingsDialog/ConnectionStatus/OK=Connected to PhotoDeck")
-    end
-
-  end, 'PhotoDeckAPI Get Gallery Attributes')
-
-  local c = f:view {
-    bind_to_object = info,
-    spacing = f:dialog_spacing(),
-
-    f:row {
-      f:static_text {
-        bind_to_object = publishSettings,
-        title = LrView.bind 'connectionStatus',
-	font = '<system/small/bold>',
-	fill_horizontal = 1
-      }
-    },
-    f:row {
-      f:static_text {
-        title = LOC("$$$/PhotoDeck/CollectionSettingsDialog/Description=Introduction:"),
-        width = LrView.share "collectionset_labelwidth",
-      },
-      f:edit_field {
-        bind_to_object = info.collectionSettings,
-        value = LrView.bind 'description',
-        width_in_chars = 60,
-        height_in_lines = 8,
-      }
-    },
-    f:row {
-      f:static_text {
-        title = LOC("$$$/PhotoDeck/CollectionSettingsDialog/GalleryDisplayStyle=Gallery Style:"),
-        width = LrView.share "collectionset_labelwidth",
-      },
-      f:static_text {
-        bind_to_object = info.collectionSettings,
-        title = LrView.bind 'galleryDisplayStyleName',
-        width_in_chars = 30,
-	fill_horizontal = 1
-      },
-      f:push_button {
-        bind_to_object = info.collectionSettings,
-        action = function() chooseGalleryDisplayStyle(publishSettings, info.collectionSettings) end,
-        enabled = LrBinding.keyIsNotNil 'galleryDisplayStyles',
-        title = LOC("$$$/PhotoDeck/CollectionSettingsDialog/ChooseGalleryDisplayStyleAction=Change Style"),
-      }
-    },
-  }
-  return c
-end
-
-publishServiceProvider.viewForCollectionSetSettings = publishServiceProvider.viewForCollectionSettings
-
+-- Update published collection
 publishServiceProvider.updateCollectionSettings = function( publishSettings, info )
   logger:trace('publishServiceProvider.updateCollectionSettings')
   PhotoDeckAPI.connect(publishSettings.apiKey, publishSettings.apiSecret, publishSettings.username, publishSettings.password)
@@ -822,6 +350,8 @@ publishServiceProvider.updateCollectionSettings = function( publishSettings, inf
   end
 end
 
+
+-- Update published collection set
 publishServiceProvider.updateCollectionSetSettings = function( publishSettings, info )
   logger:trace('publishServiceProvider.updateCollectionSetSettings')
   PhotoDeckAPI.connect(publishSettings.apiKey, publishSettings.apiSecret, publishSettings.username, publishSettings.password)
@@ -832,6 +362,8 @@ publishServiceProvider.updateCollectionSetSettings = function( publishSettings, 
   end
 end
 
+
+-- Rename published collection
 publishServiceProvider.renamePublishedCollection = function( publishSettings, info )
   logger:trace('publishServiceProvider.renamePublishedCollection')
   PhotoDeckAPI.connect(publishSettings.apiKey, publishSettings.apiSecret, publishSettings.username, publishSettings.password)
@@ -842,6 +374,8 @@ publishServiceProvider.renamePublishedCollection = function( publishSettings, in
   end
 end
 
+
+-- Reparent published collection
 publishServiceProvider.reparentPublishedCollection = function( publishSettings, info )
   logger:trace('publishServiceProvider.reparentPublishedCollection')
   PhotoDeckAPI.connect(publishSettings.apiKey, publishSettings.apiSecret, publishSettings.username, publishSettings.password)
@@ -852,6 +386,8 @@ publishServiceProvider.reparentPublishedCollection = function( publishSettings, 
   end
 end
 
+
+-- Delete published collection
 publishServiceProvider.deletePublishedCollection = function( publishSettings, info )
   logger:trace('publishServiceProvider.deletePublishedCollection')
   PhotoDeckAPI.connect(publishSettings.apiKey, publishSettings.apiSecret, publishSettings.username, publishSettings.password)
@@ -863,6 +399,8 @@ publishServiceProvider.deletePublishedCollection = function( publishSettings, in
   end
 end
 
+
+-- Go to published photo
 publishServiceProvider.goToPublishedPhoto = function( publishSettings, info )
   logger:trace('publishServiceProvider.goToPublishedPhoto')
   local catalog = LrApplication.activeCatalog()
@@ -895,4 +433,6 @@ publishServiceProvider.goToPublishedPhoto = function( publishSettings, info )
   end
 end
 
+
+-- Done
 return publishServiceProvider
